sauer: the phone on the keyboard, and four measured ways it cannot be stopped on a pad
Up is the phone in GTA IV, and up is how one walks through this menu. On the
keyboard that is solved: the trainer clears the phone control in the game's own
pad hook, and arrow up in the menu no longer brings the phone out.

On a controller it is not, and this commit records why rather than pretending.
Four attempts, each measured in the game:

  - SCRIPT_IS_USING_MOBILE_PHONE, TASK_USE_MOBILE_PHONE: no effect
  - clearing the CPad entries from the script tick: keyboard yes, pad no
  - clearing them in processPadEvent, with the history ring and all four pads:
    pad no
  - SET_PLAYER_CONTROL_FOR_TEXT_CHAT: no effect, removed again

The deciding reading came from a watcher placed at three points in one frame.
Before the clear the pad showed phone=255 frontend_up=255 kb_up=255; after the
clear, and later in the script tick, all three were zero - and the phone still
came out. The game decides inside the same call that reads the button, and the
SDK's hook only runs after it returns. There is no moment in between.

Taking the controls away entirely does stop it, and was briefly made the
default. It was rightly thrown out: a fix that costs the steering is not a fix.
The option is still in the menu, off.

What stays from the search:

  - The pad module reads the sticks now. Eight directions sit above XInput's
    16-bit button mask (so the mask is 32 bits), with two thresholds so a stick
    held near the edge does not flicker. Navigation stays on the d-pad as it was;
    the stick is there for anyone who would rather keep the phone down.
  - [Log] Inputs in sauer.ini writes down what the game sees in a frame. Off by
    default. It answered three questions tonight that would otherwise have been
    guesses.
  - The ini parser skips a byte order mark. A file saved as UTF-8 by Notepad -
    or by a PowerShell one-liner, which is how it was found - made the trainer
    complain about its own template.

136 trainer tests (62 pad, 49 config, 25 menu).

+++ src/Trainer/core/Config.h
@@ -43,7 +43,7 @@ namespace mliv
         ///
         /// One entry per chord: "L3+R3, Back" gives two, of which the first
         /// needs both sticks clicked and the second only Back.
-        std::vector<unsigned short> chords(const std::string& action) const;
+        std::vector<unsigned> chords(const std::string& action) const;
 
         bool flag(const std::string& name, bool fallback) const;
 
@@ -86,7 +86,7 @@ namespace mliv
         struct PadBinding
         {
             std::string action; ///< lower case
-            std::vector<unsigned short> chords;
+            std::vector<unsigned> chords;
         };
 
         std::vector<Entry> entries_;
